requirements.txt
python-dotenv==1.0.0
dataclasses-json==0.6.3
kafka-python==2.0.2